register progress 1
