+++ index/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 urlpatterns = [
     path('', include('cal.urls')),
     path('admin/', admin.site.urls),
-    path('register/',include('register.urls')),
+    path('',include('register.urls')),
 ]
