+++ register/templates/signup.html
@@ -41,24 +41,24 @@
 
     <div class="center2">
         <h1>Sign Up</h1>
-        <form method="post">
+        <form method="post" action="/signup/">
           <div class="txt_field">
-            <input type="text" required>
+            <input type="text" id="username" name="username" required>
             <span></span>
             <label>Username</label>
           </div>
           <div class="txt_field">
-            <input type="email" required>
+            <input type="email" id="email" name="email" required>
             <span></span>
             <label>Email</label>
           </div>
           <div class="txt_field">
-            <input type="password" required>
+            <input type="password" id="pass1" name="pass1" required>
             <span></span>
             <label>Password</label>
           </div>
           <div class="txt_field">
-            <input type="password" required>
+            <input type="password" id="pass2" name="pass2" required>
             <span></span>
             <label>Confirm Password</label>
           </div>
